Add test_equal()
I can't be completely sure that this test is correct until we support testing
arbitrary arrays in the elementwise tests (#17), although it should be. This
also corresponds to the true type promotion test for equal(), since the test
in test_type_promotion.py cannot actually test that equal() does the correct
type promotion internally.