Use keyword-only arguments for the pytest helpers functions
This makes the inscrutable helper function calls more readable, and fixes at
least one instance where the arguments were called in the wrong order.

This also changes **kw to kw={} for the function keyword arguments, so there
is no ambiguity between function keyword arguments and arguments to the helper
function.